refactor: move Profile into pop_common crate
